Refactoring sdnPSS
This commit makes sdnPSS more OF controller independent, providing an interface for other SDN controllers. If one needs to use sdnPSS with a SDN Controller other than Floodlight, he just need to implement the interface ISDNConnector. Utility functions to deal with SDNNode(s) and SDNLink(s) are also provided in this commit.