Changing OFMatch to OFRule, that now aggregate information about actions too.
This commit introduces the class OFRule, that represents an entry in an OpenFlow table.
It is basically a map of OpenFlow key,value pairs that get inserted in an OpenFlow switch
table. It is used by sdnPSS to parse an OF match and OF actions and to generate
the final Floodlight request. It also translates the entries in a format that Floodlight's
Static Flow Pusher understands.