Implementation of VLAN tagging in OFRule
This commit extends OFRule to include VLAN tagging.
It also includes lots of bug fixes and sanity checks.
This commit also includes the function reverse, which returns
a new OFRule with specifications on how to setup the "reverse"
circuit reservation (from destination to source).

Note that from now on all keys of OFRule are OpenFlow keywords.
The function floodlightEntrySet returns a set with key,values in
the format that Floodlight's Static Flow Pusher module will understand.