Implementation of implicit mode provisioning
To use implicit provisioning, the description field of the circuit reservation
needs to start with the keyword "implicit". OFRules can follow the keyword
after a comma. The URN of L1 nodes also need to encode the AID and
the router IP address using the following format:

<AID>/<ROUTER IP>

The AID should be encoded as a 48bit HEX number using dots (".") to
separate every byte. This is similar to a MAC address that uses dots to
separate each 8bit hex number instead of colons. Ip addresses should be
given in their regular format, 4 integers in the 0-255 range separated by dots.
A regular slash ("/") is used to separate AID from IP address.

The following URN is an example of a valid URN for implicit provisioning:

urn:ogf:network:domain=mydomain:node=mynode:port=1:link=1B.50.56.9B.74.C2 /123.14.0.12